Move schema store into queries
