+++ client/src/stores/queries-actions.ts
@@ -10,6 +10,7 @@ import {
   setLocalQueryText,
 } from '../utilities/localQueryText';
 import runQueryViaBatch from '../utilities/runQueryViaBatch';
+import updateCompletions from '../utilities/updateCompletions';
 import { NEW_QUERY, useQueriesStore } from './queries-store';
 
 // @ts-expect-error
@@ -415,3 +416,85 @@ export const handleChartTypeChange = (chartType: any) => {
 export const handleQuerySelectionChange = (selectedText: any) => {
   useQueriesStore.setState({ selectedText });
 };
+
+export function useShowSchema(): boolean {
+  return useQueriesStore((s) => s.showSchema);
+}
+
+export function toggleSchema() {
+  const { showSchema } = useQueriesStore.getState();
+  useQueriesStore.setState({ showSchema: !showSchema });
+}
+
+export function useSchema() {
+  return useQueriesStore((s) => s.schema);
+}
+
+export function setSchema(schema: any) {
+  useQueriesStore.setState({ schema });
+}
+
+export async function loadSchemaInfo(connectionId: string, reload?: boolean) {
+  const { showSchema, schema } = useQueriesStore.getState();
+
+  if (!schema[connectionId] || reload) {
+    setSchema({
+      ...schema,
+      [connectionId]: {
+        loading: true,
+        expanded: {},
+      },
+    });
+
+    const qs = reload ? '?reload=true' : '';
+    const json = await api.get(`/api/schema-info/${connectionId}${qs}`);
+    const { error, data } = json;
+    if (error) {
+      setSchema({
+        ...schema,
+        [connectionId]: {
+          loading: false,
+          error,
+        },
+      });
+      // If sidebar is not shown, send error notification
+      // It is otherwise shown in sidebar where schema would be
+      if (!showSchema) {
+        message.error(error);
+      }
+      return;
+    }
+    updateCompletions(data);
+
+    // Pre-expand schemas
+    const expanded: { [key: string]: boolean } = {};
+    if (data) {
+      Object.keys(data).forEach((schemaName) => {
+        expanded[schemaName] = true;
+      });
+    }
+
+    setSchema({
+      ...schema,
+      [connectionId]: {
+        loading: false,
+        schemaInfo: data,
+        error: null,
+        expanded,
+      },
+    });
+  }
+}
+
+export function toggleSchemaItem(connectionId: string, item: { id: string }) {
+  const { schema } = useQueriesStore.getState();
+  const connectionSchema = schema[connectionId];
+  const open = !connectionSchema.expanded[item.id];
+  setSchema({
+    ...schema,
+    [connectionId]: {
+      ...connectionSchema,
+      expanded: { ...connectionSchema.expanded, [item.id]: open },
+    },
+  });
+}
